+++ simulacro/src/main/java/com/riwi/simulacro/infraestructure/abstract_services/IAssignmentService.java
@@ -3,5 +3,5 @@ package com.riwi.simulacro.infraestructure.abstract_services;
 import com.riwi.simulacro.api.dto.request.AssignmentRequest;
 import com.riwi.simulacro.api.dto.response.AssignmentResponse;
 
-public interface IAssignmentService extends CRUDService<AssignmentRequest, AssignmentResponse, Integer> {
+public interface IAssignmentService extends CRUDService<AssignmentRequest, AssignmentResponse, Long> {
 }
