+++ simulacro/src/main/java/com/riwi/simulacro/infraestructure/abstract_services/ICourseService.java
@@ -3,5 +3,5 @@ package com.riwi.simulacro.infraestructure.abstract_services;
 import com.riwi.simulacro.api.dto.request.CourseRequest;
 import com.riwi.simulacro.api.dto.response.CourseResponse;
 
-public interface ICourseService extends CRUDService<CourseRequest, CourseResponse, Integer> {
+public interface ICourseService extends CRUDService<CourseRequest, CourseResponse, Long> {
 }
