+++ simulacro/src/main/java/com/riwi/simulacro/infraestructure/abstract_services/IUserService.java
@@ -3,5 +3,5 @@ package com.riwi.simulacro.infraestructure.abstract_services;
 import com.riwi.simulacro.api.dto.request.UserRequest;
 import com.riwi.simulacro.api.dto.response.UserResponse;
 
-public interface IUserService extends CRUDService<UserRequest, UserResponse, Integer> {
+public interface IUserService extends CRUDService<UserRequest, UserResponse, Long> {
 }
